requirements.txt
graphrag
networkx
matplotlib
python-docx
playwright
